feat: Usar o princípio ISP - Interface Segregation Principle (#8).
