+++ src/main/java/br/com/virtuallibrary/commons/controllers/impl/LoadController.java
@@ -1,0 +1,57 @@
+package br.com.virtuallibrary.commons.controllers.impl;
+
+import java.io.Serializable;
+
+import org.springframework.data.web.PagedResourcesAssembler;
+import org.springframework.hateoas.RepresentationModel;
+import org.springframework.hateoas.server.mvc.RepresentationModelAssemblerSupport;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+
+import br.com.virtuallibrary.commons.controllers.ILoadController;
+import br.com.virtuallibrary.commons.entities.BaseEntity;
+import br.com.virtuallibrary.commons.repositories.IBaseRepository;
+import br.com.virtuallibrary.commons.services.ILoadService;
+
+public abstract class LoadController<
+	    E extends BaseEntity,
+	    ID extends Serializable,
+	    R extends IBaseRepository<E, ID>,
+	    S extends ILoadService<E, ID, R>,
+	    M extends RepresentationModel<M>>
+    implements ILoadController<E, ID, R, S, M> {
+
+	private final S service;
+	private final RepresentationModelAssemblerSupport<E, M> modelAssembler;
+	private final PagedResourcesAssembler<E> pagedResourcesAssembler;
+
+	public LoadController(S service, PagedResourcesAssembler<E> pagedResourcesAssembler, RepresentationModelAssemblerSupport<E, M> modelAssembler) {
+		this.service = service;
+		this.modelAssembler = modelAssembler;
+		this.pagedResourcesAssembler = pagedResourcesAssembler;
+	}
+
+	@Override
+	public final S getService() {
+		return service;
+	}
+	
+	@Override
+	public PagedResourcesAssembler<E> getPagedResourcesAssembler() {
+		return pagedResourcesAssembler;
+	}
+
+	@Override
+	public RepresentationModelAssemblerSupport<E, M> getModelAssembler() {
+		return this.modelAssembler;
+	}
+	
+	@Override
+	public ResponseEntity<M> find(@PathVariable ID id) {
+		return service.findById(id) 
+			.map(modelAssembler::toModel) 
+			.map(ResponseEntity::ok) 
+			.orElse(ResponseEntity.notFound().build());
+	}
+
+}
