+++ src/main/java/br/com/virtuallibrary/rest/controllers/impl/BookController.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.web.bind.annotation.RestController;
 
-import br.com.virtuallibrary.commons.controllers.impl.BaseController;
+import br.com.virtuallibrary.commons.controllers.impl.CompleteController;
 import br.com.virtuallibrary.entity.Book;
 import br.com.virtuallibrary.repositories.BookRepository;
 import br.com.virtuallibrary.rest.controllers.IBookController;
@@ -14,9 +14,8 @@ import br.com.virtuallibrary.services.BookService;
 
 @RestController
 public class BookController
-    extends BaseController<Book, String, BookRepository, BookService, BookModel>
-    implements IBookController<Book, String, BookRepository, BookService, BookModel>{
-
+	extends CompleteController<Book, String, BookRepository, BookService, BookModel>
+    implements IBookController<Book, String, BookRepository, BookService, BookModel> {
 	@Autowired
 	public BookController(
 			BookService service,
@@ -24,5 +23,4 @@ public class BookController
 			BookModelAssembler modelAssembler) {
 		super(service, pagedResourcesAssembler, modelAssembler);
 	}
-
 }
