+++ src/main/java/br/com/virtuallibrary/services/BookService.java
@@ -4,11 +4,11 @@ import java.util.Map;
 
 import org.springframework.data.domain.Page;
 
-import br.com.virtuallibrary.commons.services.IBaseService;
+import br.com.virtuallibrary.commons.services.ICompleteService;
 import br.com.virtuallibrary.entity.Book;
 import br.com.virtuallibrary.repositories.BookRepository;
 
-public interface BookService extends IBaseService<Book, String, BookRepository> {
+public interface BookService extends ICompleteService<Book, String, BookRepository> {
 
 	Page<Book> findPaginated(int page, int size, Map<String, String> filters);
