+++ src/main/java/br/com/virtuallibrary/services/RatingService.java
@@ -2,11 +2,11 @@ package br.com.virtuallibrary.services;
 
 import java.util.List;
 
-import br.com.virtuallibrary.commons.services.IBaseService;
+import br.com.virtuallibrary.commons.services.ICompleteService;
 import br.com.virtuallibrary.entity.Rating;
 import br.com.virtuallibrary.repositories.RatingRepository;
 
-public interface RatingService extends IBaseService<Rating, String, RatingRepository> {
+public interface RatingService extends ICompleteService<Rating, String, RatingRepository> {
 
 	List<Rating> findByBookId(String bookId);
