+++ src/main/java/br/com/virtuallibrary/services/impl/RatingServiceImpl.java
@@ -5,13 +5,13 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import br.com.virtuallibrary.commons.services.impl.BaseServiceImpl;
+import br.com.virtuallibrary.commons.services.impl.CompleteServiceImpl;
 import br.com.virtuallibrary.entity.Rating;
 import br.com.virtuallibrary.repositories.RatingRepository;
 import br.com.virtuallibrary.services.RatingService;
 
 @Service
-public class RatingServiceImpl extends BaseServiceImpl<Rating, String, RatingRepository> implements RatingService {
+public class RatingServiceImpl extends CompleteServiceImpl<Rating, String, RatingRepository> implements RatingService {
 
 	@Autowired
 	public RatingServiceImpl(RatingRepository repository) {
