+++ src/test/java/br/com/virtuallibrary/othertests/RatingModelTest.java
@@ -2,6 +2,7 @@ package br.com.virtuallibrary.othertests;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.Date;
 
@@ -34,12 +35,12 @@ public class RatingModelTest {
 
 	@Test
 	public void testEqualsObjectNull() {
-		assertTrue(!entity.equals(null));
+		assertEquals(false, entity.equals(null));
 	}
 
 	@Test
 	public void testEqualsObjectNew() {
-		assertTrue(!entity.equals(new Object()));
+		assertEquals(false, entity.equals(new Object()));
 	}
 
 	@Test
@@ -50,12 +51,12 @@ public class RatingModelTest {
 	@SuppressWarnings("unlikely-arg-type")
 	@Test
 	public void testEqualsObjectNoRatingModel() {
-		assertTrue(!entity.equals(Integer.getInteger("12")));
+		assertEquals(false, entity.equals(Integer.getInteger("12")));
 	}
 
 	@Test
 	public void testEqualsNewBook() {
-		assertTrue(!entity.equals(RatingModel.builder().build()));
+		assertEquals(false, entity.equals(RatingModel.builder().build()));
 	}
 	
 	@Test
