+++ src/test/java/br/com/virtuallibrary/services/BookServicesTest.java
@@ -226,13 +226,13 @@ public class BookServicesTest {
 	@Test
 	public void testGetFilterValues() {
 		Map<String, String> filterValues = service.getFilterValues(filters);
-		assertTrue(!filterValues.containsKey(FIELD_XPTO));
+		assertEquals(false, filterValues.containsKey(FIELD_XPTO));
 	}
 	
 	@Test
 	public void testGetFilterValuesSize() {
 		Map<String, String> filterValues = service.getFilterValues(filters);
-		assertTrue(filterValues.size() == 2);
+		assertEquals(2, filterValues.size());
 	}
 	
 	@Test
